requirements.txt
numpy==1.24.3
python-dotenv==1.0.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
ultralytics==8.1.0
lapx>=0.5.2
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.26.0
pymongo==4.5.0
minio==7.1.17
python-multipart==0.0.6
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
pillow==10.0.0
requests==2.31.0
pydantic==2.4.2